requirements.txt
fastapi
uvicorn
langchain-core
langchain-openai
langchain-groq
langchain-ollama
langchain-anthropic
langchain-google-genai
langchain-mistralai
openai
llama-index
nest-asyncio
pydantic
pdfplumber
python-dotenv
python-multipart
Pillow
PyMuPDF
rapidocr-onnxruntime
